move high-level index tests to integration tests
